unset svgwrite debug mode
Debug flag is set by default so that svgwrite do some validation stuff.
If you pass paths with many segments or generally a lot of paths svgwrite
takes a lot of time for validation. The paths that svgpathtools fed are
all of correct syntax and bad attributes are ignored by almost all programs
that handled with SVG. So, setting svgwrite debug to False is good choice.