Arc.intersect() filter out bogus intersection points
Before this commit, the Arc.intersect() function returns the intersection
points of the Line with the complete ellipse implied by the Arc.

This commit filters out the intersection points that do not lie between
the `start` and `end` points of the Arc, removing incorrect intersection
points.

Note that Arc.intersect() still returns `t` parameters for the Arc that
fall outside the expected [0.0, 1.0] range.  This commit does not address
that problem.